refactor(cleanup): split cli and structural tests
