+++ src/frame_compare/cli/entry.py
@@ -1,71 +1,64 @@
 """CLI entry point for frame-compare."""
 
-# ruff: noqa: B008, E402
+# ruff: noqa: B008
 from __future__ import annotations
 
 import contextlib
-import json
 import os
 import sys
 import webbrowser
-from dataclasses import dataclass
-from enum import Enum
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, cast
-
-import tomli_w
-
-_DEFAULT_HELP_WIDTH = 200
+from typing import TYPE_CHECKING
 
 import typer
-import typer.rich_utils as typer_rich_utils
 from rich.console import Console
 
-from frame_compare.cli.errors import (
-    ExitCode,
-    format_error_console,
-    format_error_json,
-    get_exit_code,
+from frame_compare.cli.cli_helpers import (
+    FrameCompareTyperGroup,
+    handle_error,
+    prepare_toml_payload,
+    resolve_root_and_config,
+    stabilize_typer_help_width,
+    write_config_to,
 )
-from frame_compare.cli.output import print_at_a_glance, print_result_summary
-from frame_compare.config.errors import (
-    ConfigValidationError,
-    ConfigWriteError,
+from frame_compare.cli.doctor_command import (
+    doctor_report_json,
+    handle_doctor,
+    print_doctor_report,
 )
-from frame_compare.config.loader import get_default_config, load_config
-from frame_compare.config.overrides import apply_cli_overrides
-from frame_compare.config.presets import (
-    apply_preset,
-    list_presets,
-    save_preset,
+from frame_compare.cli.preset_command import (
+    handle_preset_apply,
+    handle_preset_list,
+    handle_preset_save,
 )
-from frame_compare.config.schema import (
-    ConfigSchema,
-    OverlayMode,
-    ToneCurve,
-    TonemapPreset,
-    Visibility,
+from frame_compare.cli.run_command import (
+    RunCliOptions,
+    RunCliRawArgs,
+    RunCommandDeps,
+    build_run_request_from_cli,
+    coerce_cli_choice,
+    handle_diagnose_paths,
+    handle_json_output,
+    handle_run,
 )
-from frame_compare.errors import (
-    FrameCompareError,
-    JSONValue,
+from frame_compare.cli.wizard_command import (
+    build_minimal_config,
+    handle_wizard,
+    prompt_input_dir,
+    prompt_visibility,
+    validate_config,
+    write_wizard_config_payload,
 )
+from frame_compare.config.loader import load_config
+from frame_compare.config.presets import apply_preset, list_presets, save_preset
 from frame_compare.utils.atomic_write import write_text_atomic
 from frame_compare.utils.logging import configure_logging
 
 if TYPE_CHECKING:
+    from frame_compare.config.schema import ConfigSchema
     from frame_compare.orchestration.coordinator import RunDependencies, RunRequest, RunResult
     from frame_compare.orchestration.doctor import DoctorCheck, DoctorReport
     from frame_compare.utils.progress_protocol import ProgressReporter
-
-
-def _stabilize_typer_help_width() -> None:
-    """Backfill Typer's cached Rich help width when it was imported too early."""
-    os.environ.setdefault("TERMINAL_WIDTH", str(_DEFAULT_HELP_WIDTH))
-    if typer_rich_utils.MAX_WIDTH is not None:
-        return
-    with contextlib.suppress(ValueError):
-        typer_rich_utils.MAX_WIDTH = int(os.environ["TERMINAL_WIDTH"])
 
 
 class _RunnerProxy:
@@ -89,21 +82,28 @@ def run_doctor(
     return runtime_run_doctor(checks=checks, reporter=reporter)
 
 
-from typer.core import TyperGroup
-
-
-class FrameCompareTyperGroup(TyperGroup):
-    def main(self, *args: Any, **kwargs: Any) -> Any:
-        _stabilize_typer_help_width()
-        return super().main(*args, **kwargs)
-
-
 app = typer.Typer(
     name="frame-compare",
     help="Video frame comparison tool with tonemapping and slow.pics integration.",
     no_args_is_help=False,
     cls=FrameCompareTyperGroup,
 )
+
+
+_stabilize_typer_help_width = stabilize_typer_help_width
+_prepare_toml_payload = prepare_toml_payload
+_resolve_root_and_config = resolve_root_and_config
+_doctor_report_json = doctor_report_json
+_print_doctor_report = print_doctor_report
+_prompt_input_dir = prompt_input_dir
+_prompt_visibility = prompt_visibility
+_RunCliOptions = RunCliOptions
+_build_run_request_from_cli = build_run_request_from_cli
+_coerce_cli_choice = coerce_cli_choice
+_handle_diagnose_paths = handle_diagnose_paths
+_handle_json_output = handle_json_output
+_build_minimal_config = build_minimal_config
+_validate_config = validate_config
 
 
 def _maybe_open_report(report_path: Path) -> None:
@@ -123,6 +123,14 @@ def _maybe_open_report(report_path: Path) -> None:
         return
 
 
+def _write_config_to(path: Path, config: ConfigSchema) -> None:
+    write_config_to(path, config, text_writer=write_text_atomic)
+
+
+def _write_wizard_config_payload(config_path: Path, data: dict[str, object]) -> None:
+    write_wizard_config_payload(config_path, data, text_writer=write_text_atomic)
+
+
 @app.callback(invoke_without_command=True)
 def main(ctx: typer.Context) -> None:
     """Video frame comparison tool."""
@@ -137,89 +145,6 @@ def version() -> None:
     from frame_compare import __version__
 
     typer.echo(f"frame-compare {__version__}")
-
-
-def _format_enum_expected(enum_type: type[Enum]) -> str:
-    return ", ".join(repr(member.value) for member in enum_type)
-
-
-def _coerce_cli_choice[CliChoiceT: Enum](
-    value: str | None,
-    enum_type: type[CliChoiceT],
-    loc: tuple[str, ...],
-) -> CliChoiceT | None:
-    """Convert a CLI string choice after Typer parsing so JSON errors stay structured."""
-    if value is None:
-        return None
-    try:
-        return enum_type(value)
-    except ValueError as exc:
-        expected = _format_enum_expected(enum_type)
-        raise ConfigValidationError(
-            [
-                {
-                    "type": "enum",
-                    "loc": list(loc),
-                    "msg": f"Input should be {expected}",
-                    "input": value,
-                    "ctx": {"expected": expected},
-                }
-            ]
-        ) from exc
-
-
-@dataclass(frozen=True)
-class _RunCliOptions:
-    """Typed, parsed CLI options for the run command."""
-
-    root: Path
-    config_path: Path
-    input_dir: Path | None
-    no_cache: bool
-    from_cache_only: bool
-    no_upload: bool
-    tm_preset: TonemapPreset | None
-    tm_target_nits: int | None
-    tm_curve: ToneCurve | None
-    frame_count: int | None
-    seed: int | None
-    overlay_mode: OverlayMode | None
-    skip_analysis: bool
-    skip_metadata: bool
-    skip_dovi: bool
-    force_interactive_alignment: bool
-    json_output: bool
-    no_color: bool
-    quiet: bool
-    verbose: bool
-
-
-def _build_run_request_from_cli(options: _RunCliOptions) -> RunRequest:
-    """Build the single CLI-to-runtime request mapping used by run branches."""
-    from frame_compare.orchestration.coordinator import RunRequest
-
-    return RunRequest(
-        root=options.root,
-        config_path=options.config_path,
-        input_dir=options.input_dir,
-        no_cache=options.no_cache,
-        from_cache_only=options.from_cache_only,
-        no_upload=options.no_upload,
-        tm_preset=options.tm_preset,
-        tm_target_nits=options.tm_target_nits,
-        tm_curve=options.tm_curve,
-        frame_count=options.frame_count,
-        seed=options.seed,
-        overlay_mode=options.overlay_mode,
-        skip_analysis=options.skip_analysis,
-        skip_metadata=options.skip_metadata,
-        skip_dovi=options.skip_dovi,
-        force_interactive_alignment=options.force_interactive_alignment,
-        json_output=options.json_output,
-        no_color=options.no_color,
-        quiet=options.quiet,
-        verbose=options.verbose,
-    )
 
 
 @app.command()
@@ -248,136 +173,42 @@ def run(
     verbose: bool = typer.Option(False, "--verbose", "-v"),
 ) -> None:
     resolved_root, config_path = _resolve_root_and_config(root, config)
-    effective_no_color = no_color or ("NO_COLOR" in os.environ)
-    console = Console(
-        stderr=False,
-        no_color=effective_no_color,
+    args = RunCliRawArgs(
+        resolved_root=resolved_root,
+        config_path=config_path,
+        input_dir=input_dir,
+        no_cache=no_cache,
+        from_cache_only=from_cache_only,
+        no_upload=no_upload,
+        tm_preset=tm_preset,
+        tm_target=tm_target,
+        tm_curve=tm_curve,
+        frame_count=frame_count,
+        seed=seed,
+        overlay=overlay,
+        skip_analysis=skip_analysis,
+        skip_metadata=skip_metadata,
+        skip_dovi=skip_dovi,
+        force_interactive_alignment=force_interactive_alignment,
+        json_output=json_output,
+        no_color=no_color,
+        write_config=write_config,
+        diagnose_paths=diagnose_paths,
+        quiet=quiet,
+        verbose=verbose,
     )
-    log_level = "WARNING" if quiet else ("DEBUG" if verbose else "INFO")
-    log_format = "json" if json_output else "console"
-    configure_logging(level=log_level, format=log_format)
-
-    try:
-        parsed_tm_preset = _coerce_cli_choice(tm_preset, TonemapPreset, ("color", "preset"))
-        parsed_tm_curve = _coerce_cli_choice(tm_curve, ToneCurve, ("color", "tone_curve"))
-        parsed_overlay = _coerce_cli_choice(overlay, OverlayMode, ("screenshots", "overlay_mode"))
-
-        run_options = _RunCliOptions(
-            root=resolved_root,
-            config_path=config_path,
-            input_dir=input_dir,
-            no_cache=no_cache,
-            from_cache_only=from_cache_only,
-            no_upload=no_upload,
-            tm_preset=parsed_tm_preset,
-            tm_target_nits=tm_target,
-            tm_curve=parsed_tm_curve,
-            frame_count=frame_count,
-            seed=seed,
-            overlay_mode=parsed_overlay,
-            skip_analysis=skip_analysis,
-            skip_metadata=skip_metadata,
-            skip_dovi=skip_dovi,
-            force_interactive_alignment=force_interactive_alignment,
-            json_output=json_output,
-            no_color=effective_no_color,
-            quiet=quiet,
-            verbose=verbose,
-        )
-        request = _build_run_request_from_cli(run_options)
-
-        resolved_config: ConfigSchema | None = None
-
-        def _resolve_effective_config() -> ConfigSchema:
-            return apply_cli_overrides(
-                load_config(config_path),
-                cli_args=request.cli_config_overrides(),
-            )
-
-        def _load_effective_config() -> ConfigSchema:
-            nonlocal resolved_config
-            if resolved_config is None:
-                resolved_config = _resolve_effective_config()
-            return resolved_config
-
-        if write_config:
-            _write_config_to(config_path, _load_effective_config())
-            return
-
-        if diagnose_paths:
-            _handle_diagnose_paths(resolved_root, config_path, _load_effective_config())
-            return
-
-        if not json_output and not quiet:
-            print_at_a_glance(
-                console,
-                request=request,
-                config=_load_effective_config(),
-                root=resolved_root,
-                config_path=config_path,
-            )
-
-        result = runner.run(request, dependencies=None)
-    except FrameCompareError as error:
-        if json_output:
-            typer.echo(json.dumps(format_error_json(error), sort_keys=True, separators=(",", ":")))
-            raise typer.Exit(code=int(get_exit_code(error))) from error
-        raise typer.Exit(
-            code=handle_error(error, no_color=effective_no_color, verbose=verbose)
-        ) from error
-    except KeyboardInterrupt:
-        raise typer.Exit(code=int(ExitCode.INTERRUPTED)) from None
-
-    if json_output:
-        _handle_json_output(result)
-        return
-
-    if not result.success:
-        raise typer.Exit(code=int(ExitCode.PROCESSING_ERROR))
-
-    print_result_summary(console, result=result, quiet=quiet)
-
-    if result.report_path is not None and not json_output and not quiet and sys.stdout.isatty():
-        try:
-            cfg = _resolve_effective_config()
-        except FrameCompareError:
-            # Default to opening when config cannot be reloaded so successful runs
-            # still surface the generated report in interactive sessions.
-            cfg = None
-
-        if cfg is None or cfg.report.auto_open:
-            _maybe_open_report(result.report_path)
-
-
-def _handle_diagnose_paths(resolved_root: Path, config_path: Path, config: ConfigSchema) -> None:
-    from frame_compare.orchestration.preflight import resolve_paths
-
-    workspace = resolve_paths(config, resolved_root)
-    payload = {
-        "root": str(resolved_root),
-        "config": str(config_path),
-        "input": str(workspace.input_dir),
-        "output": str(workspace.screenshots_dir),
-        "cache": str(workspace.generated_dir),
-    }
-    typer.echo(json.dumps(payload, sort_keys=True, separators=(",", ":")))
-
-
-def _handle_json_output(result: RunResult) -> None:
-    payload = {
-        "success": result.success,
-        "screenshots_dir": str(result.screenshot_dir) if result.screenshot_dir else None,
-        "slowpics_url": result.slowpics_url,
-        "report_path": str(result.report_path) if result.report_path else None,
-        "frame_count": result.frame_count,
-        "clips_processed": result.clips_processed,
-        "duration_seconds": result.duration_seconds,
-        "cache_hit": result.cache_hit,
-        "errors": list(result.errors),
-    }
-    typer.echo(json.dumps(payload, sort_keys=True, separators=(",", ":")))
-    if not result.success:
-        raise typer.Exit(code=int(ExitCode.PROCESSING_ERROR))
+    deps = RunCommandDeps(
+        runner=runner,
+        load_config=load_config,
+        write_config_to=_write_config_to,
+        handle_error=handle_error,
+        configure_logging=configure_logging,
+        console_factory=Console,
+        open_report=_maybe_open_report,
+        stdout_is_tty=sys.stdout.isatty(),
+        no_color_env_present="NO_COLOR" in os.environ,
+    )
+    handle_run(args, deps)
 
 
 @app.command()
@@ -385,56 +216,22 @@ def wizard(
     root: Path = typer.Option(".", "--root", "-r"),
     config: Path | None = typer.Option(None, "--config", "-c"),
 ) -> None:
-    """Interactive configuration wizard."""
-    defaults = get_default_config()
     resolved_root, config_path = _resolve_root_and_config(root, config)
-
-    try:
-        input_dir = _prompt_input_dir(defaults.paths.input_dir, base_dir=resolved_root)
-        auto_upload = typer.confirm(
-            "Enable slow.pics auto-upload?",
-            default=defaults.slowpics.auto_upload,
-        )
-        visibility = _prompt_visibility(defaults.slowpics.visibility)
-        delete_after_upload = typer.confirm(
-            "Delete after upload?",
-            default=defaults.slowpics.delete_after_upload,
-        )
-        tmdb_api_key = typer.prompt(
-            "TMDB API key (optional)",
-            default="",
-            hide_input=True,
-        ).strip()
-    except (KeyboardInterrupt, typer.Abort):
-        raise typer.Exit(code=int(ExitCode.INTERRUPTED)) from None
-
-    tmdb_value: str | None = tmdb_api_key if tmdb_api_key else None
-    config_data = _build_minimal_config(
-        input_dir=input_dir,
-        auto_upload=auto_upload,
-        visibility=visibility,
-        delete_after_upload=delete_after_upload,
-        tmdb_api_key=tmdb_value,
+    handle_wizard(
+        resolved_root,
+        config_path,
+        prompt_input_dir=_prompt_input_dir,
+        prompt_visibility=_prompt_visibility,
+        confirm=typer.confirm,
+        prompt_secret=typer.prompt,
+        write_payload=_write_wizard_config_payload,
+        handle_error=handle_error,
     )
-    try:
-        _validate_config(config_data)
-        _write_wizard_config_payload(config_path, config_data)
-    except FrameCompareError as error:
-        raise typer.Exit(code=handle_error(error, no_color=True, verbose=False)) from error
 
 
 @app.command()
 def doctor(json_output: bool = typer.Option(False, "--json")) -> None:
-    """Run dependency diagnostics."""
-    report = run_doctor(checks=None, reporter=None)
-    if json_output:
-        payload = _doctor_report_json(report)
-        typer.echo(json.dumps(payload, sort_keys=True, separators=(",", ":")))
-    else:
-        _print_doctor_report(report)
-
-    if report.critical_failures:
-        raise typer.Exit(code=int(ExitCode.DEPENDENCY_ERROR))
+    handle_doctor(json_output, run_doctor=run_doctor)
 
 
 preset_app = typer.Typer(name="preset", help="Manage configuration presets.", no_args_is_help=True)
@@ -446,13 +243,12 @@ def preset_list(
     root: Path = typer.Option(".", "--root", "-r"),
     config: Path | None = typer.Option(None, "--config", "-c"),
 ) -> None:
-    try:
-        resolved_root, _ = _resolve_root_and_config(root, config)
-        presets_dir = resolved_root / "config" / "presets"
-        for name in list_presets(presets_dir=presets_dir):
-            typer.echo(name)
-    except FrameCompareError as error:
-        raise typer.Exit(code=handle_error(error, no_color=True, verbose=False)) from error
+    resolved_root, _ = _resolve_root_and_config(root, config)
+    handle_preset_list(
+        resolved_root,
+        list_presets=list_presets,
+        handle_error=handle_error,
+    )
 
 
 @preset_app.command("apply")
@@ -461,14 +257,16 @@ def preset_apply(
     root: Path = typer.Option(".", "--root", "-r"),
     config: Path | None = typer.Option(None, "--config", "-c"),
 ) -> None:
-    try:
-        resolved_root, config_path = _resolve_root_and_config(root, config)
-        presets_dir = resolved_root / "config" / "presets"
-        config_data = load_config(config_path)
-        updated = apply_preset(config_data, name, presets_dir=presets_dir)
-        _write_config_to(config_path, updated)
-    except FrameCompareError as error:
-        raise typer.Exit(code=handle_error(error, no_color=True, verbose=False)) from error
+    resolved_root, config_path = _resolve_root_and_config(root, config)
+    handle_preset_apply(
+        name,
+        resolved_root,
+        config_path,
+        load_config=load_config,
+        apply_preset=apply_preset,
+        write_config_to=_write_config_to,
+        handle_error=handle_error,
+    )
 
 
 @preset_app.command("save")
@@ -477,179 +275,15 @@ def preset_save(
     root: Path = typer.Option(".", "--root", "-r"),
     config: Path | None = typer.Option(None, "--config", "-c"),
 ) -> None:
-    try:
-        resolved_root, config_path = _resolve_root_and_config(root, config)
-        presets_dir = resolved_root / "config" / "presets"
-        config_data = load_config(config_path)
-        save_preset(name, config_data, presets_dir=presets_dir)
-    except FrameCompareError as error:
-        raise typer.Exit(code=handle_error(error, no_color=True, verbose=False)) from error
-
-
-def handle_error(error: Exception, *, no_color: bool, verbose: bool) -> int:
-    """Render errors to stderr and return exit code.
-
-    Raises:
-        None.
-    """
-    if isinstance(error, FrameCompareError):
-        message = format_error_console(error, verbose=verbose)
-        if no_color:
-            typer.echo(message, err=True)
-        else:
-            console = Console(stderr=True)
-            console.print(message)
-        return int(get_exit_code(error))
-    typer.echo("Unexpected error: please report this bug.", err=True)
-    return int(ExitCode.GENERAL_ERROR)
-
-
-def _prompt_input_dir(default: str, *, base_dir: Path) -> str:
-    """Prompt for input directory and validate existence."""
-    while True:
-        value = typer.prompt("Input directory", default=default)
-        path = Path(value)
-        if not path.is_absolute():
-            path = base_dir / path
-        if path.exists() and path.is_dir():
-            return value
-        typer.echo("Input directory does not exist or is not a directory.")
-
-
-def _prompt_visibility(default: Visibility) -> str:
-    """Prompt for slow.pics visibility."""
-    allowed = {v.value for v in Visibility}
-    default_value = default.value
-    while True:
-        value = typer.prompt(
-            "slow.pics visibility (public|unlisted)",
-            default=default_value,
-        ).strip()
-        if value in allowed:
-            return value
-        typer.echo("Invalid visibility. Choose public or unlisted.")
-
-
-def _build_minimal_config(
-    *,
-    input_dir: str,
-    auto_upload: bool,
-    visibility: str,
-    delete_after_upload: bool,
-    tmdb_api_key: str | None,
-) -> dict[str, object]:
-    """Build minimal config payload for wizard output."""
-    return {
-        "paths": {"input_dir": input_dir},
-        "slowpics": {
-            "auto_upload": auto_upload,
-            "visibility": visibility,
-            "delete_after_upload": delete_after_upload,
-        },
-        "tmdb": {"api_key": tmdb_api_key},
-    }
-
-
-def _validate_config(data: dict[str, object]) -> None:
-    """Validate config data against ConfigSchema."""
-    ConfigSchema.model_validate(data)
-
-
-def _write_wizard_config_payload(config_path: Path, data: dict[str, object]) -> None:
-    """Write wizard config payload to the provided destination."""
-    try:
-        config_path.parent.mkdir(parents=True, exist_ok=True)
-        toml_text = tomli_w.dumps(_prepare_toml_payload(data))
-        write_text_atomic(config_path, toml_text, encoding="utf-8")
-    except OSError as exc:
-        raise ConfigWriteError(config_path, label="configuration file", cause=exc) from exc
-
-
-def _doctor_report_json(report: DoctorReport) -> dict[str, JSONValue]:
-    """Convert DoctorReport to JSON payload per schema."""
-    checks_payload: list[JSONValue] = []
-    for check, result in report.checks:
-        entry: dict[str, JSONValue] = {
-            "id": check.name,
-            "category": check.category,
-            "status": "pass" if result.passed else "fail",
-            "message": result.message,
-        }
-        if result.hint:
-            entry["install_hint"] = result.hint
-        if result.details:
-            entry["details"] = result.details
-        checks_payload.append(entry)
-
-    doctor_payload: dict[str, JSONValue] = {
-        "baseline_version": "R73",
-        "checks": checks_payload,
-    }
-    payload: dict[str, JSONValue] = {
-        "success": len(report.critical_failures) == 0,
-        "doctor": doctor_payload,
-    }
-    return payload
-
-
-def _resolve_root_and_config(root: Path, config: Path | None) -> tuple[Path, Path]:
-    resolved_root = Path(root).resolve()
-    if config is not None:
-        config_path = Path(config)
-        if not config_path.is_absolute():
-            config_path = (resolved_root / config_path).resolve()
-        else:
-            config_path = config_path.resolve()
-    else:
-        config_path = (resolved_root / "config" / "config.toml").resolve()
-    return resolved_root, config_path
-
-
-def _write_config_to(path: Path, config: ConfigSchema) -> None:
-    data = config.model_dump(mode="json", exclude_none=True)
-    toml_text = tomli_w.dumps(data)
-    try:
-        path.parent.mkdir(parents=True, exist_ok=True)
-        write_text_atomic(path, toml_text, encoding="utf-8")
-    except OSError as exc:
-        raise ConfigWriteError(path, label="configuration file", cause=exc) from exc
-
-
-def _prepare_toml_payload(data: dict[str, object]) -> dict[str, object]:
-    """Prepare TOML-safe payload (TOML has no null)."""
-    tmdb_section_raw = data.get("tmdb")
-    tmdb_section: dict[str, object] = {}
-    if isinstance(tmdb_section_raw, dict):
-        for k, v in cast(dict[str, object], tmdb_section_raw).items():
-            if k == "api_key":
-                if v is not None and v != "":
-                    tmdb_section["api_key"] = v
-            else:
-                tmdb_section[k] = v
-    paths_section: dict[str, object] = {}
-    slowpics_section: dict[str, object] = {}
-    paths_raw = data.get("paths")
-    slowpics_raw = data.get("slowpics")
-    if isinstance(paths_raw, dict):
-        paths_section = dict(cast(dict[str, object], paths_raw))
-    if isinstance(slowpics_raw, dict):
-        slowpics_section = dict(cast(dict[str, object], slowpics_raw))
-    payload: dict[str, object] = {
-        "paths": paths_section,
-        "slowpics": slowpics_section,
-    }
-    if tmdb_section:
-        payload["tmdb"] = tmdb_section
-    return payload
-
-
-def _print_doctor_report(report: DoctorReport) -> None:
-    """Print human-readable doctor results."""
-    for check, result in report.checks:
-        status = "PASS" if result.passed else "FAIL"
-        typer.echo(f"{status} {check.name}: {result.message}")
-        if result.hint:
-            typer.echo(f"  Hint: {result.hint}")
+    resolved_root, config_path = _resolve_root_and_config(root, config)
+    handle_preset_save(
+        name,
+        resolved_root,
+        config_path,
+        load_config=load_config,
+        save_preset=save_preset,
+        handle_error=handle_error,
+    )
 
 
 if __name__ == "__main__":
