+++ tests/windows_portable/test_windows_portable_build_scripts.py
@@ -1,0 +1,117 @@
+from __future__ import annotations
+
+import json
+import re
+from pathlib import Path
+
+from ._helpers import read_text_or_fail as _read_text_or_fail
+
+
+def test_windows_portable_bundle_launcher_sets_cwd_to_bundle_root(repo_root: Path) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert "Push-Location $bundleRoot" in build_script
+    assert "Pop-Location" in build_script
+
+
+def test_windows_portable_bundle_launcher_uses_cli_package_entry(repo_root: Path) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert "& $python -m frame_compare.cli.entry @args" in build_script
+    assert "frame_compare.cli_entry" not in build_script
+
+
+def test_windows_portable_build_creates_default_workspace_directories(repo_root: Path) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert '$bundleConfigDir = Join-Path $OutDir "config"' in build_script
+    assert '$bundleInputDir = Join-Path $OutDir "comparison_videos"' in build_script
+    assert "Ensure-Directory -Path $bundleConfigDir" in build_script
+    assert "Ensure-Directory -Path $bundleInputDir" in build_script
+
+
+def test_windows_portable_build_download_errors_name_manifest_remediation(repo_root: Path) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert "Failed to download artifact '$id' from $url" in build_script
+    assert "update $ManifestPath with a reachable URL and matching sha256" in build_script
+
+
+def test_windows_portable_ffmpeg_manifest_uses_reachable_pinned_asset_shape(
+    repo_root: Path,
+) -> None:
+    manifest_path = repo_root / "tools" / "windows_portable" / "manifest.windows-x64.json"
+    manifest = json.loads(_read_text_or_fail(manifest_path))
+    ffmpeg = next(
+        artifact for artifact in manifest["artifacts"] if artifact["id"].startswith("ffmpeg-")
+    )
+
+    assert "autobuild-2026-02-04-14-23" not in ffmpeg["url"]
+    assert "/releases/download/autobuild-" in ffmpeg["url"]
+    assert ffmpeg["url"].endswith(".zip")
+    assert len(ffmpeg["sha256"]) == 64
+    assert ffmpeg["install"]["strip_prefix"].rstrip("/") in ffmpeg["url"]
+
+
+def test_pyproject_defines_vspreview_optional_dependency(repo_root: Path) -> None:
+    pyproject_path = repo_root / "pyproject.toml"
+    pyproject = _read_text_or_fail(pyproject_path)
+    assert "[project.optional-dependencies]" in pyproject
+    assert re.search(r"vspreview\s*=\s*\[", pyproject)
+    assert re.search(r'"vspreview([^"]*)"', pyproject)
+
+
+def test_windows_portable_build_exports_vspreview_extra(repo_root: Path) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert "--extra vspreview" in build_script
+    assert "requirements.lock.txt" in build_script
+
+
+def test_windows_portable_build_has_release_public_key_gate(repo_root: Path) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert "[switch]$RequireReleasePublicKey" in build_script
+    assert "function Assert-ReleasePublicKey()" in build_script
+    assert 'Join-Path $PSScriptRoot "validate_update_public_key.ps1"' in build_script
+    assert 'Join-Path $PSScriptRoot "update_public_key.xml"' in build_script
+    assert "& $validator -PublicKeyPath $publicKey" in build_script
+    assert re.search(
+        r"function Main\(\) \{\s*if \(\$RequireReleasePublicKey\) \{\s*Assert-ReleasePublicKey",
+        build_script,
+    )
+
+
+def test_windows_portable_build_runtime_validation_checks_qt_stack(repo_root: Path) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert "import vspreview" in build_script
+    assert "import PyQt6" in build_script
+
+
+def test_windows_portable_build_writes_bundle_info_file(repo_root: Path) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert "bundle_info.json" in build_script
+    assert "requirements_lock_sha256" in build_script
+    assert "bundle_kind" in build_script
+    assert "platform" in build_script
+
+
+def test_windows_portable_build_portable_updater_launcher_fails_closed_without_exit_code(
+    repo_root: Path,
+) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert "& $updater @args" in build_script
+    assert "if ($null -eq $LASTEXITCODE) {" in build_script
+    assert "exit 1" in build_script
+    assert "if ($?)" not in build_script
+
+
+def test_windows_portable_build_copies_dist_info_licenses_when_present(repo_root: Path) -> None:
+    build_path = repo_root / "tools" / "windows_portable" / "build_portable.ps1"
+    build_script = _read_text_or_fail(build_path)
+    assert (
+        re.search(r"dist-info\\\\licenses", build_script) or "dist-info\\licenses" in build_script
+    )
